+++ inc/customizer/customizer-helpers.php
@@ -549,7 +549,10 @@ function oceanwp_cac_has_thumbnail_blog_style() {
 }
 
 function oceanwp_cac_hasnt_thumbnail_blog_style() {
-	if ( 'thumbnail-entry' == get_theme_mod( 'ocean_blog_style', 'large-entry' ) ) {
+	$style  = get_theme_mod( 'ocean_blog_style', 'large-entry' );
+	$styles = array( 'minimal-stylish', 'wide-stylish', 'thumbnail-entry' );
+
+	if ( in_array( $style, $styles, true ) ) {
 		return false;
 	} else {
 		return true;
@@ -562,6 +565,32 @@ function oceanwp_cac_has_blog_infinite_scroll() {
 	} else {
 		return false;
 	}
+}
+
+function oceanwp_cac_hasnt_default_post_header_style() {
+	$return = ( 'default' !== get_theme_mod( 'oceanwp_single_post_header_style', 'default' ) ) ? true : false;
+	return $return;
+}
+
+function oceanwp_cac_has_new_blog_style() {
+	$style  = get_theme_mod( 'ocean_blog_style', 'large-entry' );
+	$styles = array( 'minimal-stylish', 'wide-stylish' );
+
+	if ( in_array( $style, $styles, true ) ) {
+		return true;
+	} else {
+		return false;
+	}
+}
+
+function oceanwp_cac_has_top_meta_enabled() {
+	$return = ( true === get_theme_mod( 'ocean_top_entry_meta', true ) ) ? true : false;
+	return $return;
+}
+
+function oceanwp_cac_has_bottom_meta_enabled() {
+	$return = ( true === get_theme_mod( 'ocean_enable_bottom_meta', true ) ) ? true : false;
+	return $return;
 }
 
 /*-------------------------------------------------------------------------------*/