+++ inc/customizer/settings/blog.php
@@ -295,6 +295,8 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 							'large-entry'     => esc_html__( 'Large Image', 'oceanwp' ),
 							'grid-entry'      => esc_html__( 'Grid', 'oceanwp' ),
 							'thumbnail-entry' => esc_html__( 'Thumbnail', 'oceanwp' ),
+							'minimal-stylish' => esc_html__( 'Minimal Stylish', 'oceanwp' ),
+							'wide-stylish'    => esc_html__( 'Stylish Wide', 'oceanwp' )
 						),
 					)
 				)
@@ -871,6 +873,36 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 			);
 
 			/**
+			 * Blog Entries Meta Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'ocean_blog_entry_meta_style',
+				array(
+					'default'           => 'default',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'ocean_blog_entry_meta_style',
+					array(
+						'label'           => esc_html__( 'Meta Separator', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'ocean_blog_entry_meta_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'default' => esc_html__( 'Default', 'oceanwp' ),
+							'minimal' => esc_html__( 'Minimal', 'oceanwp' ),
+						),
+					)
+				)
+			);
+			/**
 			 * Blog Entries Meta Separator
 			 *
 			 * @since 2.0
@@ -897,8 +929,271 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 						'choices'  => array(
 							'default' => esc_html__( 'Default', 'oceanwp' ),
 							'modern'  => esc_html__( 'Modern', 'oceanwp' ),
+							'stylish' => esc_html__( 'Stylish', 'oceanwp' ),
 							'none'    => esc_html__( 'None', 'oceanwp' ),
 						),
+					)
+				)
+			);
+
+			/**
+			 * Enable Top Meta Styles
+			 */
+			$wp_customize->add_setting(
+				'ocean_enable_top_meta',
+				array(
+					'default'           => true,
+					'sanitize_callback' => 'oceanwp_sanitize_checkbox',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'ocean_enable_top_meta',
+					array(
+						'label'           => esc_html__( 'Enable Top Meta', 'oceanwp' ),
+						'type'            => 'checkbox',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'ocean_enable_top_meta',
+						'priority'        => 10,
+						'active_callback' => 'oceanwp_cac_has_new_blog_style',
+					)
+				)
+			);
+
+			/**
+			 * Blog Entries Top Meta for new blog entries style
+			 */
+			$wp_customize->add_setting(
+				'ocean_top_entry_meta',
+				array(
+					'default'           => apply_filters( 'ocean_blog_top_meta_default', array( 'author', 'date', 'categories', 'comments', 'mod-date', 'reading-time', 'tags' ) ),
+					'sanitize_callback' => 'oceanwp_sanitize_multi_choices',
+				)
+			);
+
+			$wp_customize->add_control(
+				new OceanWP_Customizer_Sortable_Control(
+					$wp_customize,
+					'ocean_top_entry_meta',
+					array(
+						'label'           => esc_html__( 'Top Meta', 'oceanwp' ),
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'ocean_top_entry_meta',
+						'priority'        => 10,
+						'choices'         => apply_filters(
+							'ocean_blog_top_meta_choices',
+							array(
+								'author'        => esc_html__( 'Author', 'oceanwp' ),
+								'date'          => esc_html__( 'Date', 'oceanwp' ),
+								'categories'    => esc_html__( 'Categories', 'oceanwp' ),
+								'comments'      => esc_html__( 'Comments', 'oceanwp' ),
+								'mod-date'      => esc_html__( 'Modified Date', 'oceanwp' ),
+								'reading-time'  => esc_html__( 'Reading Time', 'oceanwp' ),
+								'tags'          => esc_html__( 'Tags', 'oceanwp' ),
+							)
+						),
+						'active_callback' => 'oceanwp_cac_has_top_meta_enabled',
+					)
+				)
+			);
+
+			/**
+			 * New Blog Entries Top Meta Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_top_entry_meta_style',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'betm_style_2',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_top_entry_meta_style',
+					array(
+						'label'           => esc_html__( 'Top Meta Style', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'oceanwp_top_entry_meta_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'betm_style_2'  => esc_html__( 'Minimal', 'oceanwp' ),
+							'betm_style_3'  => esc_html__( 'Stylish', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_has_top_meta_enabled',
+					)
+				)
+			);
+
+			/**
+			 * New Blog Entries Top Meta Style Separator
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_top_entry_meta_style_separator',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'stylish',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_top_entry_meta_style_separator',
+					array(
+						'label'           => esc_html__( 'Top Meta Separator', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'oceanwp_top_entry_meta_style_separator',
+						'priority'        => 10,
+						'choices'         => array(
+							'none'     => esc_html__( 'None', 'oceanwp' ),
+							'classic'  => esc_html__( 'Classic', 'oceanwp' ),
+							'stylish'  => esc_html__( 'Stylish', 'oceanwp' ),
+							'modern'   => esc_html__( 'Modern', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_has_top_meta_enabled',
+					)
+				)
+			);
+
+			/**
+			 * Enable Bottom Meta
+			 */
+			$wp_customize->add_setting(
+				'ocean_enable_bottom_meta',
+				array(
+					'default'           => true,
+					'sanitize_callback' => 'oceanwp_sanitize_checkbox',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'ocean_enable_bottom_meta',
+					array(
+						'label'           => esc_html__( 'Enable Bottom Meta', 'oceanwp' ),
+						'type'            => 'checkbox',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'ocean_enable_bottom_meta',
+						'priority'        => 10,
+						'active_callback' => 'oceanwp_cac_has_new_blog_style',
+					)
+				)
+			);
+
+			/**
+			 * Blog Entries Bottom Meta for new blog entries style
+			 */
+			$wp_customize->add_setting(
+				'ocean_bottom_entry_meta',
+				array(
+					'default'           => apply_filters( 'ocean_blog_bottom_meta_default', array( 'author', 'date', 'categories', 'comments', 'mod-date', 'reading-time', 'tags' ) ),
+					'sanitize_callback' => 'oceanwp_sanitize_multi_choices',
+				)
+			);
+
+			$wp_customize->add_control(
+				new OceanWP_Customizer_Sortable_Control(
+					$wp_customize,
+					'ocean_bottom_entry_meta',
+					array(
+						'label'           => esc_html__( 'Bottom Meta', 'oceanwp' ),
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'ocean_bottom_entry_meta',
+						'priority'        => 10,
+						'choices'         => apply_filters(
+							'ocean_blog_bottom_meta_choices',
+							array(
+								'author'        => esc_html__( 'Author', 'oceanwp' ),
+								'date'          => esc_html__( 'Date', 'oceanwp' ),
+								'categories'    => esc_html__( 'Categories', 'oceanwp' ),
+								'comments'      => esc_html__( 'Comments', 'oceanwp' ),
+								'mod-date'      => esc_html__( 'Modified Date', 'oceanwp' ),
+								'reading-time'  => esc_html__( 'Reading Time', 'oceanwp' ),
+								'tags'          => esc_html__( 'Tags', 'oceanwp' ),
+							)
+						),
+						'active_callback' => 'oceanwp_cac_has_bottom_meta_enabled',
+					)
+				)
+			);
+
+			/**
+			 * New Blog Entries Top Meta Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_bottom_entry_meta_style',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'bebm_style_2',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_bottom_entry_meta_style',
+					array(
+						'label'           => esc_html__( 'Bottom Meta Style', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'oceanwp_bottom_entry_meta_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'bebm_style_2'  => esc_html__( 'Minimal', 'oceanwp' ),
+							'bebm_style_3'  => esc_html__( 'Stylish', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_has_bottom_meta_enabled',
+					)
+				)
+			);
+
+			/**
+			 * New Blog Entries Top Meta Style Separator
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_bottom_entry_meta_style_separator',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'stylish',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_bottom_entry_meta_style_separator',
+					array(
+						'label'           => esc_html__( 'Bottom Meta Separator', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_blog_entries',
+						'settings'        => 'oceanwp_bottom_entry_meta_style_separator',
+						'priority'        => 10,
+						'choices'         => array(
+							'none'     => esc_html__( 'None', 'oceanwp' ),
+							'classic'  => esc_html__( 'Classic', 'oceanwp' ),
+							'stylish'  => esc_html__( 'Stylish', 'oceanwp' ),
+							'modern'   => esc_html__( 'Modern', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_has_bottom_meta_enabled',
 					)
 				)
 			);
@@ -1059,6 +1354,146 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 							'sidebar-content' => esc_html__( 'Sidebar / Content', 'oceanwp' ),
 						),
 						'active_callback' => 'oceanwp_cac_has_single_post_rl_layout',
+					)
+				)
+			);
+
+			/**
+			 * Single Post Header Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_single_post_header_style',
+				array(
+					'default'           => 'default',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_single_post_header_style',
+					array(
+						'label'           => esc_html__( 'Header Style', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_single_post',
+						'settings'        => 'oceanwp_single_post_header_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'default'      => esc_html__( 'Default', 'oceanwp' ),
+							'sph_style_2'  => esc_html__( 'Intro', 'oceanwp' ),
+							'sph_style_3'  => esc_html__( 'Cover', 'oceanwp' ),
+							'sph_style_4'  => esc_html__( 'Card', 'oceanwp' ),
+							'sph_style_5'  => esc_html__( 'Card Invert', 'oceanwp' ),
+							'sph_style_6'  => esc_html__( 'Screen', 'oceanwp' ),
+							'sph_style_7'  => esc_html__( 'Screen Invert', 'oceanwp' ),
+						),
+					)
+				)
+			);
+
+			/**
+			 * Single Post Header Meta Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_single_post_header_meta_style',
+				array(
+					'default'           => 'spm_style_2',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_single_post_header_meta_style',
+					array(
+						'label'           => esc_html__( 'Header Meta Style', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_single_post',
+						'settings'        => 'oceanwp_single_post_header_meta_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'spm_style_2'  => esc_html__( 'Minimal', 'oceanwp' ),
+							'spm_style_3'  => esc_html__( 'Stylish', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_hasnt_default_post_header_style',
+					)
+				)
+			);
+
+			/**
+			 * Single Post Header Meta Style Separator
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_single_post_header_meta_separator',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'stylish',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'oceanwp_single_post_header_meta_separator',
+					array(
+						'label'           => esc_html__( 'Header Meta Separator', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_single_post',
+						'settings'        => 'oceanwp_single_post_header_meta_separator',
+						'priority'        => 10,
+						'choices'         => array(
+							'none'     => esc_html__( 'None', 'oceanwp' ),
+							'classic'  => esc_html__( 'Classic', 'oceanwp' ),
+							'stylish'  => esc_html__( 'Stylish', 'oceanwp' ),
+							'modern'   => esc_html__( 'Modern', 'oceanwp' ),
+						),
+						'active_callback' => 'oceanwp_cac_hasnt_default_post_header_style',
+					)
+				)
+			);
+
+			/**
+			 * Blog Single Header Meta
+			 */
+			$wp_customize->add_setting(
+				'oceanwp_blog_single_header_meta',
+				array(
+					'default'           => array( 'author', 'date', 'categories', 'comments', 'mod-date', 'reading-time', 'tags' ),
+					'sanitize_callback' => 'oceanwp_sanitize_multi_choices',
+				)
+			);
+
+			$wp_customize->add_control(
+				new OceanWP_Customizer_Sortable_Control(
+					$wp_customize,
+					'oceanwp_blog_single_header_meta',
+					array(
+						'label'    => esc_html__( 'Header Meta', 'oceanwp' ),
+						'section'  => 'ocean_single_post',
+						'settings' => 'oceanwp_blog_single_header_meta',
+						'priority' => 10,
+						'choices'  => apply_filters(
+							'ocean_blog_header_meta_choices',
+							array(
+								'author'        => esc_html__( 'Author', 'oceanwp' ),
+								'date'          => esc_html__( 'Date', 'oceanwp' ),
+								'categories'    => esc_html__( 'Categories', 'oceanwp' ),
+								'comments'      => esc_html__( 'Comments', 'oceanwp' ),
+								'mod-date'      => esc_html__( 'Modified Date', 'oceanwp' ),
+								'reading-time'  => esc_html__( 'Reading Time', 'oceanwp' ),
+								'tags'          => esc_html__( 'Tags', 'oceanwp' ),
+							)
+						),
+						'active_callback' => 'oceanwp_cac_hasnt_default_post_header_style',
 					)
 				)
 			);
@@ -1502,6 +1937,38 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 			);
 
 			/**
+			 * Single Post Meta Style
+			 *
+			 * @since 3.1.5
+			 */
+			$wp_customize->add_setting(
+				'ocean_blog_single_meta_style',
+				array(
+					'transport'         => 'postMessage',
+					'default'           => 'default',
+					'sanitize_callback' => 'oceanwp_sanitize_select',
+				)
+			);
+
+			$wp_customize->add_control(
+				new WP_Customize_Control(
+					$wp_customize,
+					'ocean_blog_single_meta_style',
+					array(
+						'label'           => esc_html__( 'Meta Separator', 'oceanwp' ),
+						'type'            => 'select',
+						'section'         => 'ocean_single_post',
+						'settings'        => 'ocean_blog_single_meta_style',
+						'priority'        => 10,
+						'choices'         => array(
+							'default' => esc_html__( 'Default', 'oceanwp' ),
+							'minimal' => esc_html__( 'Minimal', 'oceanwp' ),
+						),
+					)
+				)
+			);
+
+			/**
 			 * Single Post Meta Separator
 			 *
 			 * @since 2.0
@@ -1528,6 +1995,7 @@ if ( ! class_exists( 'OceanWP_Blog_Customizer' ) ) :
 						'choices'  => array(
 							'default' => esc_html__( 'Default', 'oceanwp' ),
 							'modern'  => esc_html__( 'Modern', 'oceanwp' ),
+							'stylish' => esc_html__( 'Stylish', 'oceanwp' ),
 							'none'    => esc_html__( 'None', 'oceanwp' ),
 						),
 					)