+++ changelog.md
@@ -1,5 +1,12 @@
 # OceanWP Changelog
 
+### _2022.03.08._ - 3.1.5
+
+-   **Added** New single post header style
+-   **Added** New single post meta style
+-   **Added** New blog entry style
+-   **Added** New blog entry meta style
+
 ### _2022.03.07._ - 3.1.4
 
 -   **Fixed** Logo: Retinal Logo: display issue.